feat(web): :sparkles: bootstrap web layer and user auth API
- Add FastAPI app with CORS, pagination, lifespan, and Swagger docs
- Add exception handlers and ORJSON BaseResponse envelope
- Implement auth and users v1 routes with JWT cookie flow
- Add repository layer and user schema migrations

Establishes HTTP foundation and user management for payments API.

+++ app/db/models/accounts.py
@@ -1,7 +1,7 @@
 from decimal import Decimal
 from typing import TYPE_CHECKING
 
-from sqlalchemy import ForeignKey, Integer, Numeric
+from sqlalchemy import BigInteger, ForeignKey, Numeric
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db.base import Base
@@ -16,7 +16,7 @@ class Account(Base):
 
     __tablename__ = "accounts"
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, index=True)
     user_id: Mapped[int] = mapped_column(
         ForeignKey("users.id", ondelete="CASCADE"), nullable=False
     )
